Yuhong Submission for Assignment 5
